+++ csrc/xpu/cache_ops_xpu.cpp
@@ -581,6 +581,7 @@ void gather_cached_kv(
 
 
 // scalar_t is key.scalar_type() -> half
+/*
 template <typename scalar_t, const int HD>
 void reshape_and_cache_ipexllm_kernel_fp8(
     const scalar_t* __restrict__ key,    // [num_tokens, num_heads, head_size]
@@ -664,8 +665,9 @@ void reshape_and_cache_ipexllm_kernel_fp8(
   //   value_cache[tgt_value_idx] = value[src_value_idx];
   // }
 }
+*/
 
-
+/*
 template <typename scalar_t, const int HD>
 void call_reshape_and_cache_ipexllm_kernel_fp8(
     const scalar_t* __restrict__ key, const scalar_t* __restrict__ value,
@@ -724,36 +726,37 @@ void call_reshape_and_cache_ipexllm_kernel_fp8(
         });
   });
 }
+*/
 
-void reshape_and_cache_ipexllm_fp8(torch::Tensor& key, torch::Tensor& value,
-                               torch::Tensor& key_cache,
-                               torch::Tensor& value_cache,
-                               torch::Tensor& slot_mapping,
-                               const std::string& kv_cache_dtype,
-                               const float kv_scale) {
-  int num_tokens = key.size(0);
-  int num_heads = key.size(1);
-  int head_size = key.size(2);
-  int block_size = key_cache.size(2);
-  // int x = key_cache.size(4);
-  int x = 1;
+// void reshape_and_cache_ipexllm_fp8(torch::Tensor& key, torch::Tensor& value,
+//                                torch::Tensor& key_cache,
+//                                torch::Tensor& value_cache,
+//                                torch::Tensor& slot_mapping,
+//                                const std::string& kv_cache_dtype,
+//                                const float kv_scale) {
+//   int num_tokens = key.size(0);
+//   int num_heads = key.size(1);
+//   int head_size = key.size(2);
+//   int block_size = key_cache.size(2);
+//   // int x = key_cache.size(4);
+//   int x = 1;
 
-  int key_stride = key.stride(0);
-  int value_stride = value.stride(0);
+//   int key_stride = key.stride(0);
+//   int value_stride = value.stride(0);
 
-  int key_head_stride = key.stride(1);
-  int value_head_stride = value.stride(1);
+//   int key_head_stride = key.stride(1);
+//   int value_head_stride = value.stride(1);
 
-  VLLM_XPU_DISPATCH_FLOATING_TYPES(
-      key.scalar_type(), "call_reshape_and_cache_ipexllm_kernel_fp8", [&] {
-        call_reshape_and_cache_ipexllm_kernel_fp8<scalar_t, 128>(
-            key.data_ptr<scalar_t>(), value.data_ptr<scalar_t>(),
-            key_cache.data_ptr<uint8_t>(), value_cache.data_ptr<uint8_t>(),
-            slot_mapping.data_ptr<int64_t>(), num_tokens, key_stride,
-            value_stride, key_head_stride, value_head_stride,
-            num_heads, head_size, block_size, x);
-      });
-}
+//   VLLM_XPU_DISPATCH_FLOATING_TYPES(
+//       key.scalar_type(), "call_reshape_and_cache_ipexllm_kernel_fp8", [&] {
+//         call_reshape_and_cache_ipexllm_kernel_fp8<scalar_t, 128>(
+//             key.data_ptr<scalar_t>(), value.data_ptr<scalar_t>(),
+//             key_cache.data_ptr<uint8_t>(), value_cache.data_ptr<uint8_t>(),
+//             slot_mapping.data_ptr<int64_t>(), num_tokens, key_stride,
+//             value_stride, key_head_stride, value_head_stride,
+//             num_heads, head_size, block_size, x);
+//       });
+// }
